Revert "Fixing date and time"
This reverts commit 8b2a4b184bc1cc2ed19188f47709f0a82842bbbf.
