+++ Android/app/src/main/java/hr/foi/air/fitfusion/data_classes/FirebaseManager.kt
@@ -59,37 +59,6 @@ class FirebaseManager {
         })
     }
 
-    fun addTrainingSession(
-        time: String,
-        date: String,
-        participants: String,
-        type: String,
-        userId: String,
-        callback: (Boolean, String) -> Unit
-    ) {
-        val databaseRf = FirebaseDatabase.getInstance().getReference("Training")
-
-        if (type.isNotEmpty() && participants.isNotEmpty() && time.isNotEmpty() && date.isNotEmpty()) {
-            val sessionId: String = databaseRf.push().key ?: ""
-
-            val trainingSession = TrainingModel(
-                id = sessionId,
-                date = date,
-                participants = participants,
-                state = "active",
-                time = time,
-                userId = userId,
-                type = type
-            )
-
-            databaseRf.child(sessionId).setValue(trainingSession)
-
-            callback(true, "Training session saved successfully")
-        } else {
-            callback(false, "Please fill in all the fields")
-        }
-    }
-
     fun loginUser(email: String, password: String, callback: (UserModel?, String?) -> Unit) {
         firebaseDatabase = FirebaseDatabase.getInstance()
         databaseReference = firebaseDatabase.reference.child("user")
@@ -195,8 +164,6 @@ class FirebaseManager {
         })
 
     }
-
-
 
     private val database = FirebaseDatabase.getInstance()
     private val postRef = database.getReference("Posts")
