--- Android/app/src/main/res/drawable/button_date_time.xml
@@ -1,7 +1,0 @@
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <corners android:radius="8dp" />
-    <stroke
-        android:width="2dp"
-        android:color="#000000" />
-</shape>
