+++ Android/app/src/main/res/layout/new_training_session.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <TextView
         android:id="@+id/textView"
@@ -11,9 +11,7 @@
         android:layout_height="30dp"
         android:layout_marginStart="40dp"
         android:layout_marginTop="40dp"
-        android:text="Session Type:"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"/>
+        android:text="Session Type:" />
 
     <Spinner
         android:id="@+id/spinner"
@@ -22,20 +20,14 @@
         android:layout_marginStart="40dp"
         android:layout_marginEnd="100dp"
         android:layout_marginBottom="40dp"
-        android:spinnerMode="dropdown"
-        app:layout_constraintTop_toBottomOf="@+id/textView"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
+        android:spinnerMode="dropdown" />
 
     <TextView
         android:id="@+id/number"
         android:layout_width="180dp"
         android:layout_height="30dp"
         android:layout_marginStart="40dp"
-        android:layout_marginTop="30dp"
-        android:text="Number of participants:"
-        app:layout_constraintTop_toBottomOf="@+id/spinner"
-        app:layout_constraintStart_toStartOf="parent" />
+        android:text="Number of participants:" />
 
     <EditText
         android:id="@+id/Number"
@@ -45,19 +37,14 @@
         android:layout_marginStart="40dp"
         android:ems="10"
         android:hint="Number"
-        android:inputType="number"
-        app:layout_constraintTop_toBottomOf="@+id/number"
-        app:layout_constraintStart_toStartOf="parent"/>
+        android:inputType="number" />
 
     <TextView
         android:id="@+id/time"
         android:layout_width="100dp"
         android:layout_height="30dp"
         android:layout_marginStart="40dp"
-        android:layout_marginTop="30dp"
-        android:text="Time:"
-        app:layout_constraintTop_toBottomOf="@+id/Number"
-        app:layout_constraintStart_toStartOf="parent"/>
+        android:text="Time:" />
 
     <EditText
         android:id="@+id/Time"
@@ -67,33 +54,14 @@
         android:layout_marginStart="40dp"
         android:ems="10"
         android:hint="Time"
-        android:focusable="false"
-        android:clickable="false"
-        android:inputType="none"
-        app:layout_constraintTop_toBottomOf="@+id/time"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/SelectTime"
-        android:layout_width="100dp"
-        android:layout_height="60dp"
-        android:layout_marginStart="20dp"
-        android:text="Select Time"
-        android:backgroundTint="#D3D3D3"
-        android:textColor="#111111"
-        app:layout_constraintTop_toBottomOf="@+id/time"
-        app:layout_constraintStart_toEndOf="@id/Time"
-        app:layout_constraintBottom_toBottomOf="@id/Time" />
+        android:inputType="time" />
 
     <TextView
         android:id="@+id/date"
         android:layout_width="100dp"
         android:layout_height="30dp"
         android:layout_marginStart="40dp"
-        android:layout_marginTop="30dp"
-        android:text="Date:"
-        app:layout_constraintTop_toBottomOf="@+id/Time"
-        app:layout_constraintStart_toStartOf="parent" />
+        android:text="Date:" />
 
     <EditText
         android:id="@+id/Date"
@@ -103,23 +71,7 @@
         android:layout_marginStart="40dp"
         android:ems="10"
         android:hint="Date"
-        android:focusable="false"
-        android:clickable="false"
-        android:inputType="none"
-        app:layout_constraintTop_toBottomOf="@+id/date"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/SelectDate"
-        android:layout_width="100dp"
-        android:layout_height="60dp"
-        android:layout_marginStart="20dp"
-        android:text="Select Date"
-        android:backgroundTint= "#D3D3D3"
-        android:textColor="#111111"
-        app:layout_constraintTop_toBottomOf="@+id/date"
-        app:layout_constraintStart_toEndOf="@id/Date"
-        app:layout_constraintBottom_toBottomOf="@id/Date" />
+        android:inputType="date" />
 
     <Button
         android:id="@+id/Done"
@@ -127,12 +79,8 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="20dp"
         android:layout_marginEnd="20dp"
-        android:layout_marginBottom="40dp"
-        android:layout_marginTop="40dp"
-        android:text="Done"
-        app:layout_constraintTop_toBottomOf="@+id/Date"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
+        android:layout_marginBottom="10dp"
+        android:text="Done" />
 
     <Button
         android:id="@+id/Cancel"
@@ -142,11 +90,7 @@
         android:layout_marginStart="20dp"
         android:layout_marginEnd="20dp"
         android:backgroundTint="#F44336"
-        android:layout_marginTop="20dp"
         android:text="Cancel"
-        android:textColor="#FFFFFF"
-        app:layout_constraintTop_toBottomOf="@+id/Done"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
+        android:textColor="#FFFFFF" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
